Move Treehouse-based LazyColumn into upstream module
